+++ build_runner_core/lib/src/generate/build_impl.dart
@@ -8,7 +8,8 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:build/build.dart';
-import 'package:build/internal.dart';
+// ignore: implementation_imports
+import 'package:build/src/internal.dart';
 import 'package:crypto/crypto.dart';
 import 'package:glob/glob.dart';
 import 'package:logging/logging.dart';