+++ build_test/lib/src/multi_asset_reader.dart
@@ -6,7 +6,8 @@ import 'dart:convert';
 
 import 'package:async/async.dart';
 import 'package:build/build.dart';
-import 'package:build/internal.dart';
+// ignore: implementation_imports
+import 'package:build/src/internal.dart';
 import 'package:glob/glob.dart';
 
 /// A [MultiPackageAssetReader] that delegates to multiple other asset